# docs/openapi.yaml
"openapi": "3.0.3"
"info":
  "title": "IT HelpDesk read-only API"
  "version": "1.0.0"
  "description": "Phase 4.3A. Existing session authentication only. No mutation endpoints. All known resource endpoints except session require completed password change. Unknown paths/unsupported methods return JSON 404 for authenticated users, 401 for anonymous users. See API_V1.md for examples and limitations."
"servers":
  -
    "url": "/api/v1"
"security":
  -
    "sessionCookie": []
"paths":
...
  "/tickets/{id}":
    "get":
      "operationId": "getTicket"
      "summary": "Owner or it/admin/director; other callers receive 403, missing ticket 404."
      "parameters":
        -
          "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "integer"
            "minimum": 1
            "maximum": 9007199254740991
          "description": "Positive safe integer; malformed IDs return 400."
      "responses":
        "200":
          "description": "Success"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "additionalProperties": false
                "required":
                  - "data"
                "properties":
                  "data":
                    "$ref": "#/components/schemas/TicketDetail"
        "400":
          "description": "INVALID_ID or INVALID_JSON"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Error"
        "401":
          "description": "UNAUTHENTICATED: existing Express session required"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Error"
        "403":
          "description": "FORBIDDEN or PASSWORD_CHANGE_REQUIRED"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Error"
        "404":
          "description": "NOT_FOUND: missing resource or endpoint"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Error"
        "413":
          "description": "PAYLOAD_TOO_LARGE: upstream body parser limit"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Error"
        "500":
          "description": "INTERNAL_ERROR: exception details withheld"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Error"
...
"components":
  "securitySchemes":
    "sessionCookie":
      "type": "apiKey"
      "in": "cookie"
      "name": "connect.sid"
      "description": "Obtain through existing /login form."
  "schemas":
    "Error":
      "type": "object"
      "additionalProperties": false
      "required":
        - "error"
      "properties":
        "error":
          "type": "object"
          "additionalProperties": false
          "required":
            - "code"
            - "message"
          "properties":
            "code":
              "type": "string"
            "message":
              "type": "string"
...
    "TicketDetail":
      "type": "object"
      "additionalProperties": false
      "required":
        - "id"
        - "title"
        - "description"
        - "category"
        - "priority"
        - "status"
        - "author"
        - "assignedTo"
        - "equipment"
        - "createdAt"
        - "updatedAt"
        - "firstResponseAt"
        - "startedAt"
        - "completedAt"
        - "closedAt"
        - "comments"
        - "history"
        - "attachments"
      "properties":
        "id":
          "type": "integer"
          "minimum": 1
          "maximum": 9007199254740991
        "title":
          "type": "string"
        "description":
          "type": "string"
        "category":
          "type": "string"
        "priority":
          "type": "string"
          "enum":
            - "low"
            - "normal"
            - "high"
            - "critical"
        "status":
          "type": "string"
          "enum":
            - "new"
            - "in_progress"
            - "waiting"
            - "done"
            - "closed"
        "author":
          "type": "object"
          "additionalProperties": false
          "required":
            - "id"
            - "fullName"
          "properties":
            "id":
              "type": "integer"
              "minimum": 1
              "maximum": 9007199254740991
            "fullName":
              "type": "string"
              "nullable": true
          "nullable": true
        "assignedTo":
          "type": "object"
          "additionalProperties": false
          "required":
            - "id"
            - "fullName"
          "properties":
            "id":
              "type": "integer"
              "minimum": 1
              "maximum": 9007199254740991
            "fullName":
              "type": "string"
              "nullable": true
          "nullable": true
        "equipment":
          "type": "object"
          "additionalProperties": false
          "required":
            - "id"
            - "name"
            - "assetTag"
          "properties":
            "id":
              "type": "integer"
              "minimum": 1
              "maximum": 9007199254740991
            "name":
              "type": "string"
              "nullable": true
            "assetTag":
              "type": "string"
              "nullable": true
          "nullable": true
        "createdAt":
          "type": "string"
          "nullable": true
          "description": "Existing UTC service timestamp, usually YYYY-MM-DD HH:mm:ss; null when not recorded."
        "updatedAt":
          "type": "string"
          "nullable": true
          "description": "Existing UTC service timestamp, usually YYYY-MM-DD HH:mm:ss; null when not recorded."
        "firstResponseAt":
          "type": "string"
          "nullable": true
          "description": "Existing UTC service timestamp, usually YYYY-MM-DD HH:mm:ss; null when not recorded."
        "startedAt":
          "type": "string"
          "nullable": true
          "description": "Existing UTC service timestamp, usually YYYY-MM-DD HH:mm:ss; null when not recorded."
        "completedAt":
          "type": "string"
          "nullable": true
          "description": "Existing UTC service timestamp, usually YYYY-MM-DD HH:mm:ss; null when not recorded."
        "closedAt":
          "type": "string"
          "nullable": true
          "description": "Existing UTC service timestamp, usually YYYY-MM-DD HH:mm:ss; null when not recorded."
        "comments":
          "type": "array"
          "items":
            "type": "object"
            "additionalProperties": false
            "required":
              - "id"
              - "body"
              - "author"
              - "createdAt"
            "properties":
              "id":
                "type": "integer"
                "minimum": 1
                "maximum": 9007199254740991
              "body":
                "type": "string"
              "author":
                "type": "object"
                "additionalProperties": false
                "required":
                  - "id"
                  - "fullName"
                "properties":
                  "id":
                    "type": "integer"
                    "minimum": 1
                    "maximum": 9007199254740991
                  "fullName":
                    "type": "string"
                    "nullable": true
                "nullable": true
              "createdAt":
                "type": "string"
                "nullable": true
                "description": "Existing UTC service timestamp, usually YYYY-MM-DD HH:mm:ss; null when not recorded."
        "history":
          "type": "array"
          "items":
            "type": "object"
            "additionalProperties": false
            "required":
              - "id"
              - "eventType"
              - "details"
              - "author"
              - "createdAt"
            "properties":
              "id":
                "type": "integer"
                "minimum": 1
                "maximum": 9007199254740991
              "eventType":
                "type": "string"
              "details":
                "type": "string"
                "nullable": true
              "author":
                "type": "object"
                "additionalProperties": false
                "required":
                  - "id"
                  - "fullName"
                "properties":
                  "id":
                    "type": "integer"
                    "minimum": 1
                    "maximum": 9007199254740991
                  "fullName":
                    "type": "string"
                    "nullable": true
                "nullable": true
              "createdAt":
                "type": "string"
                "nullable": true
                "description": "Existing UTC service timestamp, usually YYYY-MM-DD HH:mm:ss; null when not recorded."
        "attachments":
          "type": "array"
          "items":
            "type": "object"
            "additionalProperties": false
            "required":
              - "id"
              - "originalName"
              - "mimeType"
              - "sizeBytes"
              - "createdAt"
            "properties":
              "id":
                "type": "integer"
                "minimum": 1
                "maximum": 9007199254740991
              "originalName":
                "type": "string"
              "mimeType":
                "type": "string"
              "sizeBytes":
                "type": "integer"
              "createdAt":
                "type": "string"
                "nullable": true
                "description": "Existing UTC service timestamp, usually YYYY-MM-DD HH:mm:ss; null when not recorded."
...